The 5th task has been solved
+isUnique
+transposeIfMatrixHasNotEqualSumOfRows
*getSum
*Solve + tests

+++ libs/data_structures/matrix/matrix.c
@@ -100,8 +100,8 @@ void swapColumns(matrix *m, int j1, int j2) {
     }
 }
 
-int getSum(int *a, int n) {
-    int sum = 0;
+long long getSum(int *a, int n){
+    long long sum = 0;
     for (int i = 0; i < n; ++i)
         sum += a[i];
